Add files via upload
Perform Principal Component Analysis(PCA) on the given data set. We firstly select the number of PCA components to be as 40. We then take the number of components whose cumulative sum add up-to 100 that counts up-to 28 components. The data set is then divided into training set and test set. The ROC AUC score is found out for the original data set, PCA with 40 components and PCA with 28 components. We then perform logistic regression and get the Logit Regression Results.